+++ demo/pages/app/App.js
@@ -2,7 +2,26 @@ import { Component } from 'angular2/core';
 import { CORE_DIRECTIVES } from 'angular2/common';
 import { ROUTER_DIRECTIVES, RouteConfig, Router } from 'angular2/router';
 
-import { Home, Layout, Typography, Iconography, Color, LoadingDemo, ButtonDemo, TabsDemo, ToastDemo, CardDemo, UtilsDemo, PipesDemo, TooltipDemo, DrawerDemo, SelectDemo, SwitchDemo, DropdownDemo } from './../pages';
+import {
+    Home,
+    Layout,
+    Typography,
+    Iconography,
+    Color,
+    LoadingDemo,
+    ButtonDemo,
+    TabsDemo,
+    ToastDemo,
+    CardDemo,
+    UtilsDemo,
+    PipesDemo,
+    TooltipDemo,
+    DrawerDemo,
+    SelectDemo,
+    SwitchDemo,
+    HeaderDemo,
+    DropdownDemo
+} from './../pages';
 
 const template = require('./App.html');
 
@@ -43,8 +62,7 @@ const template = require('./App.html');
         path: '/toast',
         component: ToastDemo,
         name: 'Toast'
-    },
-    {
+    }, {
         path: '/button',
         component: ButtonDemo,
         name: 'Button'
@@ -80,6 +98,10 @@ const template = require('./App.html');
         path: '/switch',
         component: SwitchDemo,
         name: 'Switch'
+    }, {
+        path: '/header',
+        component: HeaderDemo,
+        name: 'Header'
     },
 
     // Utils
@@ -152,21 +174,10 @@ export class DemoApp {
             }, {
                 name: 'Switch',
                 path: '/switch'
+            }, {
+                name: 'Header',
+                path: '/header'
             }
-        // { name: 'Form', path: '/form' },
-        // { name: 'Dropdown', path: '/dropdowns' },
-        // { name: 'Tooltip', path: '/tooltips' },
-        // { name: 'Calendar', path: '/calendar' },
-        // { name: 'Table', path: '/table' },
-        // { name: 'Drawer', path: '/drawer' },
-        // { name: 'Switch', path: '/switch' },
-        // { name: 'Modal', path: '/modal' },
-        // { name: 'Loading', path: '/loading' },
-        // { name: 'Dragula', path: '/dragula' },
-        // { name: 'Cards', path: '/cards' },
-        // { name: 'CardExamples', path: '/card-examples' },
-        // { name: 'Chart', path: '/chart' },
-        // { name: 'Headers', path: '/headers' }
         ];
 
         this.utilRoutes = [{