+++ demo/pages/pages.js
@@ -15,3 +15,4 @@ export * from './pipes/PipesDemo';
 export * from './tooltip/TooltipDemo';
 export * from './drawer/DrawerDemo';
 export * from './switch/SwitchDemo';
+export * from './header/HeaderDemo';