+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+<a name="0.0.4"></a>
+## [0.0.4](https://github.com/bullhorn/novo-elements/compare/v0.0.3...v0.0.4) (2016-05-03)
+
+
+### Features
+
+* **drawer:** Add drawer element, closes [#4](https://github.com/bullhorn/novo-elements/issues/4)([2a16f07](https://github.com/bullhorn/novo-elements/commit/2a16f07)), closes [#4](https://github.com/bullhorn/novo-elements/issues/4)
+* **select:** Added Select element, documentation, and initial tests([f1036d2](https://github.com/bullhorn/novo-elements/commit/f1036d2))
+* **switch:** Adds switch element and documentation, closes [#16](https://github.com/bullhorn/novo-elements/issues/16)([747174b](https://github.com/bullhorn/novo-elements/commit/747174b)), closes [#16](https://github.com/bullhorn/novo-elements/issues/16)
+* **toast:** Adds toast element and documentation, closes [#18](https://github.com/bullhorn/novo-elements/issues/18)([08f2351](https://github.com/bullhorn/novo-elements/commit/08f2351)), closes [#18](https://github.com/bullhorn/novo-elements/issues/18)
+* **tooltip:** Adding tooltip feature, closes [#20](https://github.com/bullhorn/novo-elements/issues/20) ([#29](https://github.com/bullhorn/novo-elements/issues/29))([211f730](https://github.com/bullhorn/novo-elements/commit/211f730)), closes [#20](https://github.com/bullhorn/novo-elements/issues/20) [(#29](https://github.com/(/issues/29)
+
+
+
 <a name="0.0.3"></a>
 ## [0.0.3](https://github.com/bullhorn/novo-elements/compare/v0.0.2...v0.0.3) (2016-04-28)